split mapper vm integration
